update: separete icon logic in a hook
